apps/cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ralphite-cli"
version = "0.1.0"
description = "CLI surface for Ralphite"
requires-python = ">=3.12"
dependencies = [
  "typer>=0.12,<1",
  "rich>=13.9,<14",
  "pydantic>=2.8,<3",
  "pyyaml>=6.0,<7",
  "ralphite-engine",
  "ralphite-schemas"
]

[project.scripts]
ralphite = "ralphite_cli.main:app"

[tool.uv.sources]
ralphite-engine = { workspace = true }
ralphite-schemas = { workspace = true }

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
